rename and delete commented
